ai_service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=2.0.0
langchain-experimental>=0.0.49

# Gemini SDK (required by existing nexus_rag.py)
google-generativeai>=0.8.3

pydantic>=2.6.0
sqlalchemy>=2.0.25
# Database drivers
psycopg2-binary>=2.9.9  # For Postgres
supabase>=2.3.0

# File handling
python-multipart>=0.0.9
pypdf>=4.0.0

# sqlite is built-in but we might need aiosqlite for async if we go that route, keeping it simple for now
